delete test validation form and merge into LandingForm deleted files and removed comments
